feat(v2): Allow configuring babel via docusaurus.config.js
