+++ packages/docusaurus/src/babel/client-preset.ts
@@ -1,0 +1,3 @@
+import {babelPresets} from './base';
+
+module.exports = () => babelPresets(false);
