+++ packages/docusaurus/src/babel/server-preset.ts
@@ -1,0 +1,3 @@
+import {babelPresets} from './base';
+
+module.exports = () => babelPresets(true);
