+++ packages/docusaurus/src/server/config.ts
@@ -26,6 +26,7 @@ const OPTIONAL_FIELDS = [
   'scripts',
   'stylesheets',
   'tagline',
+  'configureBabel',
 ];
 
 const DEFAULT_CONFIG: {
